+++ totalsegmentator/libs.py
@@ -323,8 +323,8 @@ def download_pretrained_weights(task_id):
         weights_path = config_dir / "Dataset485_tissue_4types_1559subj"
     elif task_id == 305:
         weights_path = config_dir / "Dataset305_vertebrae_discs_1559subj"
-    elif task_id == 854:
-        weights_path = config_dir / "Dataset854_TotalSegMRI_tissue_1088subj"
+    elif task_id == 925:
+        weights_path = config_dir / "Dataset925_MRI_tissue_subset_903subj"
     elif task_id == 856:
         weights_path = config_dir / "Dataset856_TotalSegMRI_face_1088subj"
     elif task_id == 409: