+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Master
 * make `totalseg_get_modality` work with normalized intensities (images which do not have original HU values anymore)
 * replace pkg_resources for python 3.12 compatibility
+* remove torch < 2.6.0 requirement (fixed in nnU-Net)
 
 
 ## Release 2.7.0